장소 삭제 시 유저 권한 체크

+++ src/main/java/com/ft/favoritethings/spot/service/SpotService.java
@@ -1,5 +1,6 @@
 package com.ft.favoritethings.spot.service;
 
+import com.ft.favoritethings.member.entity.AccountType;
 import com.ft.favoritethings.member.entity.Member;
 import com.ft.favoritethings.spot.dto.request.SpotRequestDto;
 import com.ft.favoritethings.spot.dto.response.ResponseDto;
@@ -34,21 +35,29 @@ public class SpotService {
     }
 
     public ResponseDto<?> getAllSpots() {
+
         List<Spot> spots = spotRepository.getAllBy();
         return new ResponseDto<>(200, "Success", spots);
     }
 
     public ResponseDto<?> showSpecificSpot(Long spotId) {
+
         Optional<Spot> spotOptional = spotRepository.findById(spotId);
 
         if (spotOptional.isPresent()) {
             return ResponseDto.success(spotOptional);
         } else {
+            log.info("장소가 존재하지 않습니다.");
             return ResponseDto.fail(404, "Spot not found", "장소가 존재하지 않습니다.");
         }
     }
 
-    public ResponseDto<?> deleteSpot(Long spotId) {
+    public ResponseDto<?> deleteSpot(Long spotId, Member member) {
+
+        if (!AccountType.ROLE_ADMIN.equals(member.getAccount_type())) {
+            log.info("삭제 권한이 없는 유저입니다.");
+            return ResponseDto.fail(404, "Member is not admin",  "삭제 권한이 없는 유저입니다.");
+        }
         spotRepository.deleteById(spotId);
         return new ResponseDto<>(200, "Success", "장소 삭제");
     }
